chore(admin): finish the directory split with a rule you can actually apply
frontend/src/admin/ mixed tab-level surfaces with ordinary components, so there
was no rule a contributor could follow: a modal sat beside a tab with nothing to
distinguish them.

The rule, now true of the tree:

  admin/             entry points + the tab surfaces AdminPanel mounts (12)
  admin/components/  everything else (29)
  admin/utils/       pure logic, no JSX (21, already consistent)
  admin/hooks/       (2, already consistent)

"Is it a tab AdminPanel mounts?" is answerable without judgement, which is the
property the previous arrangement lacked.

#911 proposed a different rule -- components/ = "used by more than one tab" --
and it was rejected on the data. Only 7 of 40 files have more than one importer,
so that rule would have moved 7 files and left 31 single-use components flat,
including MfaSettingsModal, RichTextEditor and HelpPanel. The complaint the
issue opens with would have survived almost intact.

Also deletes admin/EventSelector.jsx: 186 lines, zero references anywhere in the
repo including dynamic ones -- its only mentions were its own export and doc
comment.

The issue's file counts were stale. It says 23 flat files; there were 40, and 17
of the 27 this rule relocates had already been moved. Ten remained.

Pure move. Rename similarity is 95-100%, and the only non-import lines in the
whole diff are the EventSelector deletion and two path strings in the large-file
allowlist (same 7 entries, MAX_ALLOWED unchanged -- a path update, not a
loosened gate).

make gate exits 0: 1,299 backend + 1,189 frontend tests. The build step is the
real evidence here, since a missed import path fails it.

Closes #911
