+++ frontend/src/admin/AdminPanel.jsx
@@ -8,7 +8,7 @@ import UserManagement from './UserManagement'
 import UserSettings from './UserSettings'
 import PlatformSettings from './PlatformSettings'
 import EventWizard from './EventWizard'
-import BottomNav from './BottomNav'
+import BottomNav from './components/BottomNav'
 import ContextBanner from './components/ContextBanner'
 import Breadcrumbs from './components/Breadcrumbs'
 import MfaSettingsModal from './components/MfaSettingsModal'
