+++ frontend/src/admin/EventsTab.jsx
@@ -1,11 +1,11 @@
 import { useState, useEffect, useMemo, useCallback, memo } from 'react'
 import { eventsApi, bandsApi } from '../utils/adminApi'
 import { useEventContext } from '../contexts/EventContext'
-import EventFormModal from './EventFormModal'
-import HistoricalImportModal from './HistoricalImportModal'
+import EventFormModal from './components/EventFormModal'
+import HistoricalImportModal from './components/HistoricalImportModal'
 import EventStatusBadge from './components/EventStatusBadge'
-import EmbedCodeGenerator from './EmbedCodeGenerator'
-import MetricsDashboard from './MetricsDashboard'
+import EmbedCodeGenerator from './components/EmbedCodeGenerator'
+import MetricsDashboard from './components/MetricsDashboard'
 import ArchivedEventBanner from './components/ArchivedEventBanner'
 import HelpPanel from './components/HelpPanel'
 import { Copy, Link, Ticket } from 'lucide-react'
