+++ frontend/src/admin/LineupTab.jsx
@@ -1,11 +1,11 @@
 import React, { useState, useEffect, useCallback, useMemo } from 'react'
 import { Eye, EyeOff, Ban, RotateCcw } from 'lucide-react'
 import { bandsApi, eventsApi, venuesApi } from '../utils/adminApi'
-import AnnouncementPlanningPanel from './AnnouncementPlanningPanel'
-import BandForm from './BandForm'
-import BulkActionBar from './BulkActionBar'
-import BulkPreviewModal from './BulkPreviewModal'
-import BulkBandImport from './BulkBandImport'
+import AnnouncementPlanningPanel from './components/AnnouncementPlanningPanel'
+import BandForm from './components/BandForm'
+import BulkActionBar from './components/BulkActionBar'
+import BulkPreviewModal from './components/BulkPreviewModal'
+import BulkBandImport from './components/BulkBandImport'
 import ArtistPicker from './components/ArtistPicker'
 import ConfirmDialog from '../components/ui/ConfirmDialog'
 import { parseOrigin } from '../utils/parseOrigin'
