+++ frontend/src/admin/__tests__/MetricsDashboard.test.jsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react'
 import { describe, expect, it, vi } from 'vitest'
-import MetricsDashboard from '../MetricsDashboard'
+import MetricsDashboard from '../components/MetricsDashboard'
 
 vi.mock('../../utils/adminApi', () => ({
   eventsApi: {
