+++ frontend/src/admin/components/BulkActionBar.jsx
