+++ frontend/src/admin/components/BulkPreviewModal.jsx
