+++ frontend/src/admin/components/EmbedCodeGenerator.jsx
